fixed the salsa directory code mismatch and added salsa_test helper script along with a few other code fixes

+++ src/salsa/train_and_recover.py
@@ -78,7 +78,7 @@ def get_parser():
         "--timescale", type=int, default=40, help="How fast to decay the inv sqrt lr."
     )
     parser.add_argument(
-        "--dtype", default="bfloat16", choices=["float32", "float16", "bfloat16"]
+        "--dtype", default="float16", choices=["float32", "float16", "bfloat16"]
     )
 
     # Training args
